requirements.txt
pexpect
pyalsaaudio
mps-youtube==0.2.8
pyro5
flask
flask-cors
